(#70): const부분 각 영역으로 분리
