+++ src/entities/tour/consts.ts
@@ -1,1 +1,13 @@
 export const NUM_OF_ROWS = 10;
+
+export const TOUR_TYPE = {
+  '': '없음',
+  12: '관광지',
+  14: '문화시설',
+  15: '행사/공연/축제',
+  25: '여행코스',
+  28: '레포츠',
+  32: '숙박',
+  38: '쇼핑',
+  39: '음식점',
+};
