+++ src/features/aroundTourist/const.ts
@@ -54,15 +54,3 @@ export const markerImageMap = {
   '38': touristContentSVG.ShoppingPin,
   '39': touristContentSVG.RestaurantPin,
 } as const;
-
-export const TOUR_TYPE = {
-  '': '없음',
-  12: '관광지',
-  14: '문화시설',
-  15: '행사/공연/축제',
-  25: '여행코스',
-  28: '레포츠',
-  32: '숙박',
-  38: '쇼핑',
-  39: '음식점',
-};
