+++ src/features/aroundTourist/index.ts
@@ -1,2 +1,3 @@
 export * from './lib';
 export * from './ui';
+export * from './const';
