--- src/pages/const/TMAP.ts
@@ -1,7 +1,0 @@
-export const TMAP_APP_KEY = import.meta.env.VITE_TMAP_APP_KEY; // 여기에 실제 API 키 입력
-export const TRAFFIC = {
-  none: '#888888',
-  normal: '#24aa24',
-  slowly: '#ffc107',
-  delay: '#dc3545',
-} as const;
