--- src/pages/const/index.ts
@@ -1,4 +1,0 @@
-export * as MARKER from './MARKER';
-export * as TMAP from './TMAP';
-export * as FOLLOW from './FOLLOW';
-export * as TRANSPORT from './TRANSPORT';
